[loading] Fix ordering of requires
The config file needs to be required after oboe has been required,
since it changes the behavior of the config object that's created in
oboe.

This whole setup between oboe and oboe_fu should be revisited. Most of
the API code belongs in the oboe module, but for now we're getting it to
work with this rewrite.