[api/ar] Fix bug in log_event API call
The check for whether kvs were passed to log_event was incorrect,
failing on nil.

This bug was triggered by the active record instrumentation. Since
'opts' was only being set if tracing was disabled, referencing 'opts'
outside of the if would evaluate to nil, which was then passed into
log_event.