Seems like when there are user documents, it fails
I think the existing test is providing flags in a way Moped doesn't know
how to handle, but so far it's been OK because the collection has been
empty. It's failing now, but I think it's failing because I inserted
some users, and now it has to update some documents and can't.