[profiling] version with omitted timestamps from other threads
In this version time spent in another threads gets added as omitted to
each of the profiled threads, see:
https://docs.google.com/document/d/1f0Ui9jjZwZ5GmbePygWLB1bbpc5DDAKSfewVFc8LLY8/edit#

but to account for all the time spent in other threads the timestamps
need to be updated from threads that are not tracing/profiling and that
would be a big overhead.

This implementation only accounts for time spent in other threads
that are profiling and therefor is not very useful.

AO-12790